Changed SamParser API methods from canAccept() and accept(...) to canParse() and parse(...).
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@2586 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ src/org/jcvi/jillion/trace/fastq/FastqFileParser.java
@@ -245,7 +245,7 @@ public abstract class FastqFileParser implements FastqParser{
             }
         }
         //now parse the qualities
-        int expectedQualities =  (int)sequenceBuilder.length();
+        int expectedQualities =  sequenceBuilder.length();
         
         //clear builders
         sequenceBuilder.setLength(0);
